Level class: refactor level and exceptions relating it
