+++ src/test/java/seedu/address/testutil/TuitionClassBuilder.java
@@ -5,7 +5,7 @@ import java.util.Set;
 
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tuitionclass.Day;
-import seedu.address.model.tuitionclass.Level;
+import seedu.address.model.level.Level;
 import seedu.address.model.tuitionclass.Name;
 import seedu.address.model.tuitionclass.Subject;
 import seedu.address.model.tuitionclass.Time;
